Update 7.General Tips for School.md
Added a section where it provides suggestions to those who came in with a lot of credits. 

I came in with nearly 80 credits, and those were the different options I weighed. I believe that giving these ideas out to those who are in similar situation as I am may be beneficial

By Eric Liu